[REFACTOR] Temp folder names
