+++ src/react-app/tempcomponents/ProteinNode.tsx
