+++ src/react-app/tempcomponents/SimulationDrawer.tsx
