+++ src/react-app/tempcomponents/tempedges/index.ts
